requirements.txt
PyYAML==6.0.1
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
flask==3.0.0
jinja2==3.1.2
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
streamlit==1.29.0
plotly==5.17.0
streamlit-agraph==0.0.45
streamlit-option-menu==0.3.6
graphviz==0.20.1
pillow==10.1.0
requests==2.31.0
reportlab==4.0.7
seaborn==0.12.2
kaleido==0.2.1
